+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 - Added Numeric => `many?`
 - Added Numeric => `one?`
 - Added Numeric => `none?` to `zero?` alias
+- Added Numeric => `to_range`
 
 ## [2.0.2] - 2021-07-22
 ### Changed